Improving the usability of the version system and making generic
Originally this was 'my next project' because I decided to make a
boilerplate, so cleaning up some of the left over things.

Also, so you don't have to modify the Version.cmake to set the version,
converted it over to a function that you call, if you want to use
versioning.  Generates in the form of:

<major>.<minor>.<patch>-<git commit id>